Proven additions of the --extra and --sender flags.
Holding off on the -l (list options) and -V or --version as they
require a little more working around the standard flag input
behaviour.  Both functions are presently supported through the -f and
-h flags anyway.

Raised version number and will merge with master.  Conversion of all
options to use relevant -s and/or -e flags will progress gradually
following this.

A minor fix to one of the field options was also made.